Add "cppbenchmark" library from Conan and benchmark the Compressor and Decompressor
