+++ CMakeLists.txt
@@ -21,12 +21,15 @@ include(FetchContent)
 include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
 conan_basic_setup(TARGETS)
 
+# #############################################################################
+# Storage setup
 set(EUGENE_STORAGE src/core/storage)
 set(LibEugeneCore_SRC
         Position.h PageCache.h Page.h PageCache.cpp Page.cpp
-        Storage.h
+        Storage.h Logger.h
         btree/Btree.h btree/BtreePrinter.h btree/Btree.cpp
-        compression/Compressor.h compression/Decompressor.h compression/Compression.cpp)
+        compression/Compressor.h compression/Decompressor.h
+        compression/TestCompression.cpp)
 list(TRANSFORM LibEugeneCore_SRC PREPEND "${EUGENE_STORAGE}/")
 
 add_library(LibEugeneCore "${LibEugeneCore_SRC}")
@@ -36,8 +39,10 @@ target_link_libraries(LibEugeneCore
         CONAN_PKG::fmt
         CONAN_PKG::ms-gsl
         CONAN_PKG::outcome
+        CONAN_PKG::catch2
         CONAN_PKG::spdlog
-        CONAN_PKG::catch2)
+        CONAN_PKG::cppbenchmark)
+
 
 if (DEFINED ENV{EUGENE_BUILD_TESTS})
     macro(eugene_test component path_prefix)
@@ -51,14 +56,14 @@ if (DEFINED ENV{EUGENE_BUILD_TESTS})
     message("-- Building core unit tests")
 
     add_library(TestMain src/core/Main.cpp)
-    target_link_libraries(TestMain CONAN_PKG::catch2)
+    target_link_libraries(TestMain LibEugeneCore)
 
-    eugene_test(Compression ${EUGENE_STORAGE}/compression)
+    eugene_test(TestCompression ${EUGENE_STORAGE}/compression)
     eugene_test(Btree ${EUGENE_STORAGE}/btree)
     eugene_test(Page ${EUGENE_STORAGE})
     eugene_test(PageCache ${EUGENE_STORAGE})
 
-    # add executable for compression/Benchmark.cpp
-    add_executable(benchmark ${EUGENE_STORAGE}/compression/Benchmark.cpp)
-    target_link_libraries(benchmark PUBLIC LibEugeneCore TestMain)
+    # add executable for compression/benchmarks/BenchmarkCompressor.cpp
+    add_executable(benchmark ${EUGENE_STORAGE}/compression/BenchmarkCompression.cpp)
+    target_link_libraries(benchmark PUBLIC LibEugeneCore)
 endif ()
