+++ conanfile.txt
@@ -4,6 +4,7 @@ catch2/2.13.7
 fmt/8.0.1
 spdlog/1.9.2
 ms-gsl/3.1.0
+cppbenchmark/cci.20201029
 
 [generators]
 cmake
